+++ crates/omnix-common/src/lib.rs
@@ -1,2 +1,3 @@
 //! Omnix library crate
+pub mod check;
 pub mod logging;